# docs/swagger.yaml
basePath: /
definitions:
  auth.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/auth.UserInfo'
    type: object
  auth.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  auth.GoogleLoginRequest:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    required:
    - email
    - external_id
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.OTPResponse:
    properties:
      expires_in:
        type: integer
      message:
        type: string
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        type: string
    required:
    - email
    - password
    type: object
  auth.SendOTPRequest:
    properties:
      identifier:
        description: email or phone
        type: string
      method:
        enum:
        - email
        - sms
        type: string
    required:
    - identifier
    - method
    type: object
  auth.TelegramVerifyRequest:
    properties:
      init_data:
        description: Telegram WebApp initData
        type: string
    required:
    - init_data
    type: object
  auth.UserInfo:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      phone_number:
        type: string
      telegram_username:
        type: string
    type: object
  auth.VerifyOTPRequest:
    properties:
      code:
        type: string
      identifier:
        type: string
    required:
    - code
    - identifier
    type: object
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: Authentication Service API for Nabi application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nabi Auth API
  version: latest
paths:
  /auth/.well-known/jwks.json:
    get:
      description: Get the public key in JWKS format
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Get Public Key
      tags:
      - auth
  /auth/google/login:
    post:
      consumes:
      - application/json
      description: Login using Google token details directly
      parameters:
      - description: Google Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Login with Google (Direct)
      tags:
      - oauth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the current session/token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/oauth2/google:
    get:
      description: Redirect to Google OAuth login page
      responses:
        "307":
          description: Redirect to Google
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Initiate Google OAuth
      tags:
      - oauth
  /auth/oauth2/google/callback:
    get:
      description: Handle Google OAuth callback
      parameters:
      - description: Authorization Code
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Google OAuth Callback
      tags:
      - oauth
  /auth/otp/send:
    post:
      consumes:
      - application/json
      description: Send One-Time Password to email or phone
      parameters:
      - description: Send OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.OTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Send OTP
      tags:
      - auth
  /auth/otp/verify:
    post:
      consumes:
      - application/json
      description: Verify One-Time Password
      parameters:
      - description: Verify OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Verify OTP
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email or phone number and password
      parameters:
      - description: Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/telegram/verify:
    post:
      consumes:
      - application/json
      description: Verify Telegram WebApp login data
      parameters:
      - description: Telegram Verify Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.TelegramVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Verify Telegram Login
      tags:
      - oauth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"